% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dbw.R
\name{print.dbw}
\alias{print.dbw}
\title{Print the distribution balancing weighting estimation results}
\usage{
\method{print}{dbw}(x, ...)
}
\arguments{
\item{x}{an object of class “dbw”, usually, a result of a call to \code{\link{dbw}}.}

\item{...}{additional arguments to be passed to print.}
}
\value{
No retrun value, called for side effects.
}
\description{
Prints a fitted \code{dbw} object.
}
\seealso{
\code{\link{dbw}}, \code{\link[base]{print}}
}
\author{
Hiroto Katsumata
}
