% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std.R
\name{std}
\alias{std}
\title{Standardize covariates for propensity score estimation}
\usage{
std(x)
}
\arguments{
\item{x}{a data frame (or one that can be coerced to that class)
containing the covariates for the propensity score estimation.}
}
\value{
A data frame of the standardized covariates, where the intercept term
is coerced to be a vector of 1s if included.
}
\description{
Standardizes covariates for propensity score estimation before the
distribution balancing weighting \code{dbw} when using the regularization.
}
\examples{
# For examples see example(dbw)
}
\seealso{
\code{\link{dbw}}
}
\author{
Hiroto Katsumata
}
